requirements.txt
opencv-python~=4.11.0.86
numpy~=2.1.1
ultralytics~=8.3.80
torch~=2.6.0
torchvision~=0.21.0
pillow~=11.1.0
matplotlib~=3.10.1